requirements.txt
# Core dependencies
dlib>=19.24.0
Pillow>=10.0.0
numpy>=1.24.0

# Optional: for YAML config support
PyYAML>=6.0

